# perf(crypto): simplify and speed up Address.Compare