--- a/tm2/pkg/bft/consensus/common_test.go
+++ b/tm2/pkg/bft/consensus/common_test.go
@@ -128,7 +128,8 @@ func (vss ValidatorStubsByAddress) Len() int {
 }
 
 func (vss ValidatorStubsByAddress) Less(i, j int) bool {
-	return vss[i].GetPubKey().Address().Compare(vss[j].GetPubKey().Address()) == -1
+	vai := vss[i].GetPubKey().Address()
+	return vai.Compare(vss[j].GetPubKey().Address()) == -1
 }
 
 func (vss ValidatorStubsByAddress) Swap(i, j int) {