--- a/tm2/pkg/bft/types/validator_set_test.go
+++ b/tm2/pkg/bft/types/validator_set_test.go
@@ -1319,7 +1319,7 @@ func (valz validatorsByPriority) Less(i, j int) bool {
 	if valz[i].ProposerPriority > valz[j].ProposerPriority {
 		return false
 	}
-	return valz[i].Address.Compare(valz[j].Address) < 0
+	return valz[i].Address.Compare(&valz[j].Address) < 0
 }
 
 func (valz validatorsByPriority) Swap(i, j int) {