Takes advantage of knowledge that Address is simply an array derived from []byte and Go's rules allow us to address each with [:] to find []byte and pass those into bytes.Compare instead of the unnecessarily inefficient make, copy. crypto.Address is very fundamental and a building block to Gno and all of cryptocurrencies, thus making its operations much more efficient is necessary.
While here also made Address.Compare take a pointer receiver which reduces the need to make a stack copy of the receiver and with that we have this speed up while making the code more succinct and direct and more idiomatic

```shell
$ benchstat before.txt after.txt
name              old time/op    new time/op    delta
AddressCompare-8    79.5ns ± 1%    33.3ns ± 8%  -58.17%  (p=0.000 n=10+9)

name              old alloc/op   new alloc/op   delta
AddressCompare-8     8.00B ± 0%     8.00B ± 0%     ~     (all equal)

name              old allocs/op  new allocs/op  delta
AddressCompare-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```

Fixes #4129