@thehowl 2025-04-14
I don't know if this matters enough; it seems to be used in some functions in Validator/ValidatorSet to matter enough. I'll let someone else take a second look first.