@moul 2025-04-15
I'm also skeptical about how much it matters. We should have not only "per function benchmarks" but also flame graphs that show how frequently calls are made in real usage.

These flame graphs could become shared assets rather than something we expect with each pull request.

I'll open an issue to explore if someone has a good idea for generating flame graphs in the CI. We don't need a history; ideally, we just want a few generated flame graphs that match a recent version of the code or a pull request for convenience.