@moul 2025-04-15
-> https://github.com/gnolang/gno/issues/4151