@odeke-em 2025-04-15
One thing though to objectively acknowledge and note is that the prior code was really extraneous and my change isn't even an "optimization" but really how the code should have been written for idiomatic Go the proper way, it just happens to make the code much more efficient and faster. 

Skepticism over performance is usually more pertinent when one is being clever and the performance checks will later show, Address comparison is pretty important and quite heavily used.